web-server/requirements.txt
Flask==2.1.2
numpy==1.23.0
pymongo==4.1.1
Werkzeug==2.1.2
requests==2.28.1
pika==1.3.0
python-dotenv
scikit-learn==1.2.2
kneed
typing_extensions